src/django/src/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.7
django-filters==0.2.1
djangorestframework==3.11.0
idna==2.9
lxml==4.5.1
mysqlclient==1.4.6
Pygments==2.6.1
PyMySQL==0.9.3
pytz==2020.1
requests==2.23.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
